requirements.txt
alembic==0.9.5
appdev.py==0.0.1
autoenv==1.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Kivy==1.10.1.dev0
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.6
marshmallow-sqlalchemy==0.13.1
mysql-connector==2.1.4
MySQL-python==1.2.5
nose==1.3.7
numpy==1.14.3
python-dateutil==2.6.1
python-editor==1.0.3
PyX==0.14.1
redis==2.10.6
requests==2.18.4
six==1.10.0
SQLAlchemy==1.1.15
urllib3==1.22
Werkzeug==0.12.2
